requirements.txt
pyyaml >= 5.4.1
beautifulsoup4 >= 4.9.3